feat(quant): 添加持仓策略并修改交易接口
- 新增 FixedAmountPositionStrategy 固定买卖金额持仓策略
- 新增 FixedVolumePositionStrategy 固定买卖量持仓策略
- 新增 FullPositionStrategy 满仓策略
- 修改 Account 接口将 volume 参数从 long 改为 int
- 修改 PositionStrategy 接口增加 BarSeries 参数
- 在 Trader 中启用 positionStrategy 和 tradeFeeStrategy 的非空验证
- 在测试类中使用 FullPositionStrategy 替代 null 值

+++ clever-task-example/src/main/java/org/clever/quant/Account.java
@@ -38,7 +38,7 @@ public interface Account {
      * @param fee    手续费(交易成本)
      * @return 如果没有交易就返回 null
      */
-    TradeLog enter(long barIdx, double price, long volume, double fee);
+    TradeLog enter(long barIdx, double price, int volume, double fee);
 
     /**
      * 平仓
@@ -49,5 +49,5 @@ public interface Account {
      * @param fee    手续费(交易成本)
      * @return 如果没有交易就返回 null
      */
-    TradeLog exit(long barIdx, double price, long volume, double fee);
+    TradeLog exit(long barIdx, double price, int volume, double fee);
 }
