+++ clever-task-example/src/main/java/org/clever/quant/account/SimulationAccount.java
@@ -25,12 +25,12 @@ public class SimulationAccount extends AbstractAccount {
     }
 
     @Override
-    public TradeLog enter(long barIdx, double price, long volume, double fee) {
+    public TradeLog enter(long barIdx, double price, int volume, double fee) {
         return null;
     }
 
     @Override
-    public TradeLog exit(long barIdx, double price, long volume, double fee) {
+    public TradeLog exit(long barIdx, double price, int volume, double fee) {
         return null;
     }
 }
